% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_markdown_files.R
\name{create_project_markdown_files}
\alias{create_project_markdown_files}
\title{Create Project Markdown}
\usage{
create_project_markdown_files()
}
\value{
Rmd file
}
\description{
This function creates project Rmarkdown files from project templates
}
\examples{
\dontrun{
create_project_markdown_files()
 }
}
