% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_to_ddm.R
\name{dd_to_ddm}
\alias{dd_to_ddm}
\title{Convert decimal degrees (DD) to degrees decimal minutes (DDM).
This function only accepts coordinates north of the equator and west of the prime meridian.}
\usage{
dd_to_ddm(dd_coord)
}
\arguments{
\item{dd_coord}{Coordinate in DD}
}
\value{
The given coordinate converted to DDM
}
\description{
Convert decimal degrees (DD) to degrees decimal minutes (DDM).
This function only accepts coordinates north of the equator and west of the prime meridian.
}
\examples{

boat_coords <- data.frame(lat = 44.58509, lon = -80.93900)
boat_coords$dd_lat <- dd_to_ddm(boat_coords$lat)
boat_coords$dd_lon <- dd_to_ddm(boat_coords$lon)

}
